% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/longitudinalcascadefunction.R
\name{long.cascade}
\alias{long.cascade}
\title{Longitudinal cascade generation function}
\usage{
long.cascade(df, stages.order, groups.order = NA, death.indicator = NA,
  retention.indicator = NA, censorship.indicator = NA,
  allow.sub.lines = FALSE, allow.skips = FALSE, x.axis.max = 365)
}
\arguments{
\item{df}{(required) The main dataframe input parameter. The data frame needs at least the following fields:
"ID": (required) Either a numerical or string-based individual identifier, indicating every person in the dataset
"date": (required) Date-formatted date son which the event / stage occurred
"stage": (required) String indicating the stage achieved by the individual on the specified date. Stages must match the string in the stages.order parameter. Additonal events may be included in the "stage" category, including death, loss to follow up, and interstage events defined in the other parameters.
"group": (optional) Strings indicating any relevent groups of data.}

\item{stages.order}{(required) stages.order is the parameter which defines the events to be considered in the main cascade and their order. This is a vector of strings matching items in the "Stage" column of the main data frame, e.g. c("Stage 1","Stage 2","Stage 3").}

\item{groups.order}{(optional) This is a vector of groups, matching the "group" column of the}
}
\description{
This package generates a longitudinal casade, including a graphical representation. This takes a long-formatted list of stage-by-stage events and transforms it into a longitudinal cascade, correcting the orders of events.
}
\examples{
long.cascade.chart()
}
\keyword{cascade}
\keyword{longitudinal}
\keyword{survival}
